# contracts/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: BhoomiSetu AI API Specification
  description: REST API specification for BhoomiSetu AI - Early Detection of Land Acquisition Delays
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local Development Server
paths:
  /auth/login:
    post:
      summary: Authenticate user and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      summary: Retrieve current authenticated user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
...
  /cases/{id}:
    get:
      summary: Get case details with parcels, tasks, and risk status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Case details
...
  /documents/upload:
    post:
      summary: Upload document (PDF/Image) for OCR extraction
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                case_id:
                  type: integer
                parcel_id:
                  type: integer
                document_type:
                  type: string
      responses:
        '201':
          description: Document uploaded and OCR processed
  /documents/{id}/ocr-result:
    get:
      summary: Get OCR extracted fields and discrepancy report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OCR result
  /documents/{id}/verify:
    put:
      summary: Update document verification status (Officer action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Document verified
  /predictions/case/{case_id}:
    post:
      summary: Trigger on-demand AI risk prediction for a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: AI delay prediction result
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT